Remove notion of behaviors (see desc); finish StateReplicator
Behaviors were bad because they oftentimes forced janky scene tree structures and were hard to integrate with helpers. They are being replaced with simple scripts at the root of an object that contain references (hard or dynamic) to the children. This keeps the benefit of keeping the tree structure in an immediately obvious spot while removing a bunch of restrictions. Root nodes also reduce the number of semi-expensive node lookups when compared to the old `RootNode.GetBehavior` calls.